requirements.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
bentoml==1.3.8
bitarray==2.9.2
black==24.2.0
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
cattrs==23.1.2
cbor2==5.6.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
circus==0.18.0
click==8.1.8
click-option-group==0.5.6
cloudpickle==3.1.0
colab-xterm==0.2.0
coloredlogs==15.0.1
comm==0.2.2
cryptography==43.0.1
Cython==3.0.11
cytoolz==0.12.3
dataclasses-json==0.6.7
dataclassy==0.11.1
debugpy==1.8.7
decorator==5.1.1
deepmerge==2.0
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
eip712==0.2.4
eval_type_backport==0.2.2
execnet==2.0.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastembed-gpu==0.5.1
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask_wtf==1.2.2
flatbuffers==25.2.10
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2025.2.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
httpcore==1.0.7
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.0.1
hypothesis==6.27.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwt==1.3.1
kaitaistruct==0.10
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-mistralai==0.2.7
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langsmith==0.1.134
lark==1.2.2
lazy-object-proxy==1.10.0
loguru==0.7.3
lru-dict==1.2.0
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistralai==1.5.0
mistune==3.0.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.3
nvidia-cudnn-cu12==9.1.0.70
nvidia-ml-py==11.525.150
openai==1.64.0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parsimonious==0.9.0
parso==0.8.4
pathspec==0.12.1
pdfminer==20191125
pexpect==4.9.0
pillow==10.4.0
pip-requirements-parser==32.0.1
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.10.1
prometheus_client==0.21.0
prompt-toolkit==3.0.43
propcache==0.3.0
protobuf==5.29.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-solc-ast==1.2.10
py-solc-x==1.1.1
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic-settings==2.5.2
pydantic_core==2.27.2
Pygments==2.17.2
pygments-lexer-solidity==0.7.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==5.3.0
pyshadow==0.0.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2025.1
pyunormalize==15.1.0
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.13.2
referencing==0.33.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
rlp==4.0.0
rpds-py==0.18.0
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.25.0
selenium-wire==5.1.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.3.1
setuptools==75.1.0
simple-di==0.1.5
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.42.2
streamlit-pdf-viewer==0.0.21
striprtf==0.0.26
sympy==1.13.3
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli_w==1.1.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uv==0.4.20
uvicorn==0.31.1
vvm==0.1.0
watchfiles==0.24.0
wcwidth==0.2.13
web3==6.15.1
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.42.0
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.17.0
zstandard==0.23.0
